Membuat aplikasi daftar planet tata surya dengan Fragment dan ViewModel
