+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/reyhan/labtkpl/views/MainActivity.java
@@ -1,0 +1,33 @@
+package id.ac.ui.cs.mobileprogramming.reyhan.labtkpl.views;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
+
+import id.ac.ui.cs.mobileprogramming.reyhan.labtkpl.R;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        setTitle("Solar System");
+
+        // Create new fragment and transaction
+        Fragment planetListFragment = new PlanetListFragment();
+        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+
+        // Replace whatever is in the fragment_container view with this fragment,
+        // and add the transaction to the back stack
+        transaction.add(R.id.fragment1, planetListFragment);
+
+        // Commit the transaction
+        transaction.commit();
+    }
+}
